chore: Add RootVisual Uno-specific logic to XamlIsland as well
