+++ src/Uno.UI/UI/Xaml/Internal/IRootElement.cs
@@ -1,0 +1,6 @@
+﻿namespace Uno.UI.Xaml.Core;
+
+internal interface IRootElement
+{
+	UnoRootElementLogic RootElementLogic { get; }
+}
