+++ src/Uno.UI/UI/Xaml/Internal/Islands/XamlIsland.cs
@@ -1,23 +1,28 @@
 ﻿#nullable enable
 
 using Uno.UI.Xaml.Controls;
+using Uno.UI.Xaml.Core;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using WinUICoreServices = global::Uno.UI.Xaml.Core.CoreServices;
 
 namespace Uno.UI.Xaml.Islands;
 
-internal partial class XamlIsland : Panel
+internal partial class XamlIsland : Panel, IRootElement
 {
 	private readonly ContentManager _contentManager;
+	private readonly UnoRootElementLogic _rootElementLogic;
 
 	public XamlIsland()
 	{
 		_contentManager = new(this, false);
+		_rootElementLogic = new(this);
 		InitializeRoot(WinUICoreServices.Instance);
 	}
 
 	internal ContentManager ContentManager => _contentManager;
+
+	UnoRootElementLogic IRootElement.RootElementLogic => _rootElementLogic;
 
 	public UIElement? Content
 	{
